+++ src/Harvest/Middleware/Greenleaf.php
@@ -13,15 +13,24 @@ use DecodeLabs\Greenleaf\Context;
 use DecodeLabs\Greenleaf\Dispatcher;
 use DecodeLabs\Greenleaf\RouteNotFoundException;
 use DecodeLabs\Greenleaf\Route\Hit;
-use Psr\Http\Message\ResponseInterface as Response;
-use Psr\Http\Message\ServerRequestInterface as Request;
-use Psr\Http\Server\MiddlewareInterface as Middleware;
-use Psr\Http\Server\RequestHandlerInterface as Handler;
+use DecodeLabs\Harvest\Middleware as HarvestMiddleware;
+use DecodeLabs\Harvest\MiddlewareGroup;
+use Psr\Http\Message\ResponseInterface as PsrResponse;
+use Psr\Http\Message\ServerRequestInterface as PsrRequest;
+use Psr\Http\Server\RequestHandlerInterface as PsrHandler;
 
 class Greenleaf implements
     Dispatcher,
-    Middleware
+    HarvestMiddleware
 {
+    public MiddlewareGroup $group {
+        get => MiddlewareGroup::Generator;
+    }
+
+    public int $priority {
+        get => 0;
+    }
+
     protected Context $context;
 
     public function __construct(
@@ -34,8 +43,8 @@ class Greenleaf implements
      * Begin stage stack navigation
      */
     public function handle(
-        Request $request
-    ): Response {
+        PsrRequest $request
+    ): PsrResponse {
         $hit = $this->getHit($request);
 
         return $hit->getRoute()->handleIn(
@@ -49,9 +58,9 @@ class Greenleaf implements
      * Handle request
      */
     public function process(
-        Request $request,
-        Handler $next
-    ): Response {
+        PsrRequest $request,
+        PsrHandler $next
+    ): PsrResponse {
         try {
             $hit = $this->getHit($request);
 
@@ -69,7 +78,7 @@ class Greenleaf implements
      * Perform routing
      */
     protected function getHit(
-        Request &$request
+        PsrRequest &$request
     ): Hit {
         $hit = $this->context->matchIn($request, true);
         $request = $request->withAttribute('route', $hit->getRoute());