+++ tests/AnalyzeJsonApiTrait.php
@@ -12,16 +12,13 @@ namespace DecodeLabs\Greenleaf\Tests;
 use DecodeLabs\Greenleaf\Action;
 use DecodeLabs\Greenleaf\Action\JsonApiTrait;
 use DecodeLabs\Harvest\Response\Json as JsonResponse;
-use Psr\Http\Message\ResponseInterface as Response;
-use Psr\Http\Message\ServerRequestInterface as Request;
-use DecodeLabs\Singularity\Url\Leaf as LeafUrl;
-
+use Psr\Http\Message\ResponseInterface as PsrResponse;
 
 class AnalyzeJsonApiTrait implements Action
 {
     use JsonApiTrait;
 
-    public function get(): Response {
+    public function get(): PsrResponse {
         return new JsonResponse(['foo' => 'bar']);
     }
 }